Use virtual lines to show output
This commit makes nvim-bqn use virtual lines feature of neovim 0.6.0
show output of running BQN expressions.

The set of available command has changed, first we have the following
commands to eval BQN expressions:

    :EvalBQNTillLine
    :EvalBQNRange
    :EvalBQNFile

and then a set of commands to clear output:

    :ClearBQNAfterLine
    :ClearBQNRange
    :ClearBQNFile

There's one change to the mappings is that `<CR>` is not mapped to
`:EvalBQNTillLine` and thus it means it evaluates everything from the
start of the buffer till the line the cursor is at. This behaviour is
useful when you bindings defined above the cursor.

When rendering output the highlight group is used either `bqnoutok` or
`bqnouterr` depending on if the interpreter returned an error. By
default those groups are configured as:

    hi link bqnoutok Comment
    hi link bqnouterr Error