Redraw after rendering output
This work arounds another bug with neovim virtual lines: when we update
output for a line and the new output contains less lines than previously
neovim behaves weird (the previous output is not cleared properly).

The redraw command at the end of an update fixes this so neovim properly
redraws the grid.